Update CMake source file
- employ modern CMake and raise minimum required version of CMake to 3.13
- use `find_package` to resolve dependencies
  - use installed dependencies if there are already present on the system
- use `add_subdirectory` call or CMake `ExternalProject` module for bundled dependencies
  - change handling of Protobuf, Kokkos, kokkos-kernels, Catch2, htslib, and libvgio
- define target `libpairg` for PairG library
- define target `pairg` for PairG command-line tool
- avoid modifying `CMAKE_CXX_FLAGS` since it is set by CMake for different build types
  - set compiler standard flag through `target_compile_feature` for `libpairg` target
- different build type for `pairg` and `tests` targets
- rename `BUILD_TESTS` to `BUILD_TESTING` which is more conventional and set it off by
  default

**Note**:
This commit makes `libvgio` a **required** dependency which was bundled before. This is
because libvgio does not handle its dependencies properly (e.g. libhandlegraph).
Therefore, it cannot be simply bundled by calling `add_subdirectory`. Considering that
handling libvgio's dependencies should be done in its own package, it seems reasonable
to ask users to have it installed as a requirement. This commit, however, does not
remove `libvgio` submodule in case this change needs to be reverted to make it bundled
again. The required procedure for bundling VGio is included but commented out because of
aforementioned issue in VGio build script.